backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
firebase-admin==6.3.0
requests==2.31.0
PyJWT==2.8.0
gunicorn==21.2.0
APScheduler==3.10.4
google-generativeai>=0.8.0
geopy==2.4.1
google-cloud-storage==2.14.0
PyPDF2==3.0.1
openpyxl==3.1.2
